fix(islamic): respect selected snooze duration in prayer overlay
The snooze button was always using 5 minutes regardless of the duration
selected by the user (5/10/15/30 minutes). Thread the selected duration
through the callback chain so the timer uses the correct interval.

+++ Crest/Views/Prayer/PrayerOverlayWindow.swift
@@ -3,9 +3,9 @@ import SwiftUI
 
 final class PrayerOverlayWindow: NSPanel {
     private let onDismissAction: () -> Void
-    private let onSnoozeAction: () -> Void
+    private let onSnoozeAction: (Int) -> Void
 
-    init(prayer: Prayer, prayerTime: Date, onDismiss: @escaping () -> Void, onSnooze: @escaping () -> Void) {
+    init(prayer: Prayer, prayerTime: Date, onDismiss: @escaping () -> Void, onSnooze: @escaping (Int) -> Void) {
         self.onDismissAction = onDismiss
         self.onSnoozeAction = onSnooze
